+++ test/browser/features/fixtures/packages/vue-router/src/views/ContactView.vue
@@ -1,0 +1,12 @@
+<script setup>
+import { onMounted } from 'vue'
+
+onMounted(() => { document.title = 'Contact 1'; })
+</script>
+
+<template>
+  <div>
+    <h1>Contact 1</h1>
+  </div>
+</template>
+