+++ CHANGELOG.md
@@ -12,6 +12,11 @@
 - (delivery-fetch) Added `@bugsnag/delivery-fetch-performance` package [#281](https://github.com/bugsnag/bugsnag-js-performance/pull/281)
 - (react-native) Create simple generator for span and trace IDs [#282](https://github.com/bugsnag/bugsnag-js-performance/pull/262)
 
+### Added
+
+- (delivery-fetch) Added `@bugsnag/delivery-fetch-performance` package [#281](https://github.com/bugsnag/bugsnag-js-performance/pull/281)
+- (react-native) Create simple generator for span and trace IDs [#282](https://github.com/bugsnag/bugsnag-js-performance/pull/262)
+
 ### Fixed
 
 - (browser) Do not retry delivery for oversized payloads when connection fails [#280](https://github.com/bugsnag/bugsnag-js-performance/pull/280)